@@ -15,29 +15,28 @@
     }
   ],
   "require": {
-    "php": "^7.0",
-    "zendframework/zend-expressive": "^2.0",
+    "php": "^7.2",
+    "ext-json": "*",
     "zendframework/zend-stdlib": "^3.1",
     "zendframework/zend-servicemanager": "^3.3",
     "zendframework/zend-http": "^2.5",
     "reputation-vip/queue-client": "^2.0",
     "opis/closure": "^2.1.0",
     "rollun-com/rollun-utils": "^4.2",
-    "rollun-com/rollun-actionrender": "^6.0",
     "rollun-com/rollun-logger": "^4.0",
-    "http-interop/http-middleware": "0.4.1",
-    "zendframework/zend-json": "^3.1"
+    "zendframework/zend-json": "^3.1",
+    "zendframework/zend-stratigility": "^3.0",
+    "zendframework/zend-diactoros": "^2.0"
   },
   "require-dev": {
-    "zendframework/zend-component-installer": "^1.0 ",
-    "zendframework/zend-config-aggregator": "^1.0",
-    "zendframework/zend-expressive-helpers": "^4.0",
-    "zendframework/zend-expressive-fastroute": "^2.0",
-    "zendframework/zend-expressive-zendviewrenderer": "^1.4",
-    "phpunit/phpunit": "^5.7.15",
-    "squizlabs/php_codesniffer": "^2.8.1",
-    "zfcampus/zf-development-mode": "^3.1",
-    "filp/whoops": "^2.1.7"
+    "zendframework/zend-config-aggregator": "^1.1",
+    "zendframework/zend-expressive": "^3.2",
+    "zendframework/zend-expressive-helpers": "^5.1",
+    "zendframework/zend-expressive-fastroute": "^3.0",
+    "symfony/dotenv": "^4.2",
+    "phpunit/phpunit": "^7.0",
+    "squizlabs/php_codesniffer": "^3.0",
+    "filp/whoops": "^2.3"
   },
   "autoload": {
     "psr-4": {
@@ -50,22 +49,11 @@
     }
   },
   "scripts": {
-    "post-create-project-cmd": [
-      "@development-enable"
-    ],
-    "development-disable": "zf-development-mode disable",
-    "development-enable": "zf-development-mode enable",
-    "development-status": "zf-development-mode status",
-    "check": [
-      "@cs-check",
-      "@test"
-    ],
-    "clear-config-cache": "php bin/clear-config-cache.php",
-    "cs-check": "phpcs",
-    "cs-fix": "phpcbf",
-    "serve": "php -S 0.0.0.0:8080 -t public public/index.php",
     "test": "phpunit --colors=always",
     "test-coverage": "phpunit --colors=always --coverage-clover clover.xml",
-    "upload-coverage": "coveralls -v"
+    "upload-coverage": "coveralls -v",
+    "code-sniffer": "phpcs .",
+    "code-beautiful": "phpcbf .",
+    "server": "php -S 0.0.0.0:8000 -t public public/index.php"
   }
 }
