@@ -1,24 +1,22 @@
 <?php
+/**
+ * @copyright Copyright © 2014 Rollun LC (http://rollun.com/)
+ * @license LICENSE.md New BSD License
+ */
 
 namespace rollun\test\callback\Callback;
 
-require_once './src/Callback/src/Callback/Example/CallMe.php';
-
-use rollun\callback\Callback\Callback;
+use PHPUnit\Framework\TestCase;
 use rollun\callback\Callback\Example\CallMe;
-use rollun\callback\Callback\Interruptor\Http;
-use rollun\callback\Callback\Interruptor\Process;
-use rollun\callback\Callback\Promiser;
-use rollun\dic\InsideConstruct;
-use rollun\installer\Command;
+use rollun\callback\Callback\Interrupter\Process;
 
 /**
- * Generated by PHPUnit_SkeletonGenerator on 2016-09-20 at 12:54:48.
+ * Class CallbackTestDataProvider
+ * @package rollun\test\callback\Callback
  */
-class CallbackTestDataProvider extends \PHPUnit_Framework_TestCase
+class CallbackTestDataProvider extends TestCase
 {
-
-    public function provider_mainType()
+    public function providerMainType()
     {
         $stdObject = (object)['prop' => 'Hello '];
 
@@ -28,8 +26,8 @@ class CallbackTestDataProvider extends \PHPUnit_Framework_TestCase
                 'class_parents',
                 self::class,
                 [
-                    'PHPUnit_Framework_TestCase' => "PHPUnit_Framework_TestCase",
-                    'PHPUnit_Framework_Assert' => "PHPUnit_Framework_Assert"
+                    'PHPUnit\Framework\TestCase' => "PHPUnit\Framework\TestCase",
+                    'PHPUnit\Framework\Assert' => "PHPUnit\Framework\Assert"
                 ]
             ],
             //closure
@@ -79,7 +77,7 @@ class CallbackTestDataProvider extends \PHPUnit_Framework_TestCase
         );
     }
 
-    public function provider_multiplexerType()
+    public function providerMultiplexerType()
     {
         $stdObject = (object)['prop' => 'Hello '];
         //function
@@ -120,14 +118,6 @@ class CallbackTestDataProvider extends \PHPUnit_Framework_TestCase
     public function providerTickerType()
     {
         return [
-            /*
-             * [
-             *     $tickerCallback,
-             *     $ticksCount,
-             *     $tickDuration,
-             *     $delayMC,
-             * ]
-             */
             [
                 function ($val) {
                     file_put_contents($val, microtime(true) . "\n", FILE_APPEND);
