@@ -1,24 +1,36 @@
 <?php
+/**
+ * @copyright Copyright © 2014 Rollun LC (http://rollun.com/)
+ * @license LICENSE.md New BSD License
+ */
 
-use Zend\ConfigAggregator\ArrayProvider;
+use Symfony\Component\Dotenv\Dotenv;
 use Zend\ConfigAggregator\ConfigAggregator;
 use Zend\ConfigAggregator\PhpFileProvider;
 
-// To enable or disable caching, set the `ConfigAggregator::ENABLE_CACHE` boolean in
-// `config/autoload/local.php`.
-$cacheConfig = [
-    'config_cache_path' => 'data/config-cache.php',
-];
+// Make environment variables stored in .env accessible via getenv(), $_ENV or $_SERVER.
+(new Dotenv())->load('.env');
+
+// Determine application environment ('dev' or 'prod').
+$appEnv = getenv('APP_ENV');
 
 $aggregator = new ConfigAggregator([
-    // Include cache configuration
-    new ArrayProvider($cacheConfig),
-    new \Zend\Expressive\ConfigProvider(),
-    new \Zend\Expressive\Router\ConfigProvider(),
-    new \rollun\actionrender\ConfigProvider(),
+    \Zend\Expressive\ConfigProvider::class,
+    \Zend\HttpHandlerRunner\ConfigProvider::class,
+    \Zend\Log\ConfigProvider::class,
+    \Zend\Db\ConfigProvider::class,
+    \Zend\Expressive\Router\ConfigProvider::class,
+    \Zend\Expressive\Helper\ConfigProvider::class,
+    \Zend\Cache\ConfigProvider::class,
+    \Zend\Mail\ConfigProvider::class,
+    \Zend\Validator\ConfigProvider::class,
+    \Zend\Expressive\Router\FastRouteRouter\ConfigProvider::class,
+
+    // Rollun providers
+    \rollun\callback\ConfigProvider::class,
+    \rollun\logger\ConfigProvider::class,
 
     // Default App module config
-
     // Load application config in a pre-defined order in such a way that local settings
     // overwrite global settings. (Loaded as first to last):
     //   - `global.php`
@@ -27,8 +39,12 @@ $aggregator = new ConfigAggregator([
     //   - `*.local.php`
     new PhpFileProvider('config/autoload/{{,*.}global,{,*.}local}.php'),
 
-    // Load development config if it exists
-    new PhpFileProvider('config/development.config.php'),
-], $cacheConfig['config_cache_path']);
+    // Load application config according to environment:
+    //   - `global.dev.php`,   `global.test.php`,   `prod.global.prod.php`
+    //   - `*.global.dev.php`, `*.global.test.php`, `*.prod.global.prod.php`
+    //   - `local.dev.php`,    `local.test.php`,     `prod.local.prod.php`
+    //   - `*.local.dev.php`,  `*.local.test.php`,  `*.prod.local.prod.php`
+    new PhpFileProvider(realpath(__DIR__) . "/autoload/{{,*.}global.{$appEnv},{,*.}local.{$appEnv}}.php"),
+]);
 
 return $aggregator->getMergedConfig();
