@@ -1,9 +1,14 @@
 <?php
+/**
+ * @copyright Copyright © 2014 Rollun LC (http://rollun.com/)
+ * @license LICENSE.md New BSD License
+ */
 
+use Psr\Container\ContainerInterface;
 use Zend\Expressive\Application;
-use Zend\Expressive\Handler\NotFoundHandler;
 use Zend\Expressive\Helper\ServerUrlMiddleware;
 use Zend\Expressive\Helper\UrlHelperMiddleware;
+use Zend\Expressive\MiddlewareFactory;
 use Zend\Expressive\Router\Middleware\DispatchMiddleware;
 use Zend\Expressive\Router\Middleware\ImplicitHeadMiddleware;
 use Zend\Expressive\Router\Middleware\ImplicitOptionsMiddleware;
@@ -12,42 +17,47 @@ use Zend\Stratigility\Middleware\ErrorHandler;
 
 /**
  * Setup middleware pipeline:
+ * @param Application $app
+ * @param MiddlewareFactory $factory
+ * @param ContainerInterface $container
+ * @return void
  */
-/** @var $app Application */
-// The error handler should be the first (most outer) middleware to catch
-// all Exceptions.
-$app->pipe(ErrorHandler::class);
-$app->pipe(ServerUrlMiddleware::class);
+return function (Application $app, MiddlewareFactory $factory, ContainerInterface $container): void {
+    // The error handler should be the first (most outer) middleware to catch
+    // all Exceptions.
+    $app->pipe(ErrorHandler::class);
+    $app->pipe(ServerUrlMiddleware::class);
 
-// Pipe more middleware here that you want to execute on every request:
-// - bootstrapping
-// - pre-conditions
-// - modifications to outgoing responses
-//
-// Piped Middleware may be either callables or service names. Middleware may
-// also be passed as an array; each item in the array must resolve to
-// middleware eventually (i.e., callable or service name).
-//
-// Middleware can be attached to specific paths, allowing you to mix and match
-// applications under a common domain.  The handlers in each middleware
-// attached this way will see a URI with the MATCHED PATH SEGMENT REMOVED!!!
-//
-// - $app->pipe('/api', $apiMiddleware);
-// - $app->pipe('/docs', $apiDocMiddleware);
-// - $app->pipe('/files', $filesMiddleware);
+    // Pipe more middleware here that you want to execute on every request:
+    // - bootstrapping
+    // - pre-conditions
+    // - modifications to outgoing responses
+    //
+    // Piped Middleware may be either callables or service names. Middleware may
+    // also be passed as an array; each item in the array must resolve to
+    // middleware eventually (i.e., callable or service name).
+    //
+    // Middleware can be attached to specific paths, allowing you to mix and match
+    // applications under a common domain.  The handlers in each middleware
+    // attached this way will see a URI with the MATCHED PATH SEGMENT REMOVED!!!
+    //
+    // - $app->pipe('/api', $apiMiddleware);
+    // - $app->pipe('/docs', $apiDocMiddleware);
+    // - $app->pipe('/files', $filesMiddleware);
 
-// Register the routing middleware in the middleware pipeline
-$app->pipe(RouteMiddleware::class);
-$app->pipe(ImplicitHeadMiddleware::class);
-$app->pipe(ImplicitOptionsMiddleware::class);
-$app->pipe(UrlHelperMiddleware::class);
+    // Register the routing middleware in the middleware pipeline
+    $app->pipe(RouteMiddleware::class);
+    $app->pipe(ImplicitHeadMiddleware::class);
+    $app->pipe(ImplicitOptionsMiddleware::class);
+    $app->pipe(UrlHelperMiddleware::class);
 
-// Add more middleware here that needs to introspect the routing results; this
-// might include:
-//
-// - route-based authentication
-// - route-based validation
-// - etc.
+    // Add more middleware here that needs to introspect the routing results; this
+    // might include:
+    //
+    // - route-based authentication
+    // - route-based validation
+    // - etc.
 
-// Register the dispatch middleware in the middleware pipeline
-$app->pipe(DispatchMiddleware::class);
+    // Register the dispatch middleware in the middleware pipeline
+    $app->pipe(DispatchMiddleware::class);
+};
