@@ -1,7 +1,16 @@
 <?php
-// Delegate static file requests back to the PHP built-in webserver
-use rollun\logger\LifeCycleToken;
 
+declare(strict_types=1);
+
+use rollun\dic\InsideConstruct;
+use rollun\logger\LifeCycleToken;
+use Zend\Expressive\Application;
+use Zend\Expressive\MiddlewareFactory;
+use Zend\ServiceManager\ServiceManager;
+
+error_reporting(E_ALL ^ E_USER_DEPRECATED);
+
+// Delegate static file requests back to the PHP built-in webserver
 if (php_sapi_name() === 'cli-server'
     && is_file(__DIR__ . parse_url($_SERVER['REQUEST_URI'], PHP_URL_PATH))
 ) {
@@ -10,30 +19,32 @@ if (php_sapi_name() === 'cli-server'
 
 chdir(dirname(__DIR__));
 require 'vendor/autoload.php';
-require_once 'config/env_configurator.php';
-
 
 /**
  * Self-called anonymous function that creates its own scope and keep the global namespace clean.
  */
 call_user_func(function () {
-    /** @var \Interop\Container\ContainerInterface $container */
+    /** @var ServiceManager $container */
     $container = require 'config/container.php';
-    \rollun\dic\InsideConstruct::setContainer($container);
-    //inject token to container
-    $lifeCycleToke = LifeCycleToken::generateToken();
-    if(apache_request_headers() && array_key_exists("LifeCycleToken", apache_request_headers())) {
-        $lifeCycleToke->unserialize(apache_request_headers()["LifeCycleToken"]);
-    }
-    $container->setService(LifeCycleToken::class, $lifeCycleToke);
+    InsideConstruct::setContainer($container);
 
-    /** @var \Zend\Expressive\Application $app */
-    $app = $container->get(\Zend\Expressive\Application::class);
+    /** @var Application $app */
+    $app = $container->get(Application::class);
+    $factory = $container->get(MiddlewareFactory::class);
 
     // Import programmatic/declarative middleware pipeline and routing
     // configuration statements
-    require 'config/pipeline.php';
-    require 'config/routes.php';
+    (require 'config/pipeline.php')($app, $factory, $container);
+    (require 'config/routes.php')($app, $factory, $container);
+
+    // Init lifecycle token
+    $lifeCycleToken = LifeCycleToken::generateToken();
+
+    if (LifeCycleToken::getAllHeaders() && array_key_exists("LifeCycleToken", LifeCycleToken::getAllHeaders())) {
+        $lifeCycleToken->unserialize(LifeCycleToken::getAllHeaders()["LifeCycleToken"]);
+    }
+
+    $container->setService(LifeCycleToken::class, $lifeCycleToken);
 
     $app->run();
 });
