@@ -1,17 +1,13 @@
 <?php
-
 /**
- * Zaboy lib (http://zaboy.org/lib/)
- *
- * @copyright  Zaboychenko Andrey
- * @license http://opensource.org/licenses/gpl-license.php GNU Public License
+ * @copyright Copyright © 2014 Rollun LC (http://rollun.com/)
+ * @license LICENSE.md New BSD License
  */
 
 namespace rollun\callback\Callback\Example;
 
 class CallMe
 {
-
     public function __invoke($val)
     {
         return 'Hello ' . $val;
@@ -26,5 +22,4 @@ class CallMe
     {
         return 'Hello ' . $val;
     }
-
 }
