@@ -1,5 +1,8 @@
 <?php
-
+/**
+ * @copyright Copyright © 2014 Rollun LC (http://rollun.com/)
+ * @license LICENSE.md New BSD License
+ */
 
 namespace rollun\callback\Middleware;
 
@@ -17,9 +20,10 @@ class GetParamsResolver extends AbstractParamsResolver
      * @param ServerRequestInterface $request
      * @return ServerRequestInterface
      */
-    protected function resolveParams(ServerRequestInterface $request)
+    protected function resolveParams(ServerRequestInterface $request): ServerRequestInterface
     {
         $value = $request->getQueryParams();
+
         return $request->withAttribute(static::ATTRIBUTE_WEBHOOK_VALUE, $value);
     }
 }
