@@ -34,7 +34,8 @@ class QueueInstaller extends InstallerAbstract
         return $config;
     }
 
-    public function getName() {
+    public function getName()
+    {
 
         return static::class;
     }
