@@ -1,61 +1,50 @@
 <?php
+/**
+ * @copyright Copyright © 2014 Rollun LC (http://rollun.com/)
+ * @license LICENSE.md New BSD License
+ */
 
 namespace rollun\test\callback\Callback;
 
-use rollun\callback\Callback\Callback;
-use rollun\test\callback\Callback\CallbackTestDataProvider;
-use rollun\promise\Promise\Promise;
-use rollun\dic\InsideConstruct;
+use rollun\callback\Callback\SerializedCallback;
 
 /**
- * Generated by PHPUnit_SkeletonGenerator on 2016-09-20 at 12:54:48.
+ * Class CallbackTest
+ * @package rollun\test\callback\Callback
  */
 class CallbackTest extends CallbackTestDataProvider
 {
-
     /**
-     * @covers \rollun\callback\Callback\Callback::__invoke
-     * @dataProvider provider_mainType()
+     * @covers \rollun\callback\Callback\SerializedCallback::__invoke
+     * @dataProvider providerMainType
      */
-    public function test__invoke($callable, $val, $expected)
+    public function testInvoke($callable, $val, $expected)
     {
-        $callback = new Callback($callable);
-        /* @var $callback Callback */
+        $callback = new SerializedCallback($callable);
+        /* @var $callback SerializedCallback */
         $this->assertEquals($expected, $callback($val));
     }
 
     /**
-     * @covers \rollun\callback\Callback\Callback::__sleep
-     * @dataProvider provider_mainType()
+     * @covers \rollun\callback\Callback\SerializedCallback::__sleep
+     * @dataProvider providerMainType()
      */
-    public function test__sleep($callable, $val, $expected)
+    public function testSleep($callable, $val, $expected)
     {
-        $callback = new Callback($callable);
-        /* @var $callback Callback */
+        $callback = new SerializedCallback($callable);
+        /* @var $callback SerializedCallback */
         $this->assertEquals('array', gettype($callback->__sleep()));
     }
 
     /**
-     * @covers \rollun\callback\Callback\Callback::__wakeup
-     * @dataProvider provider_mainType()
+     * @covers \rollun\callback\Callback\SerializedCallback::__wakeup
+     * @dataProvider providerMainType()
      */
-    public function test__wakeup($callable, $val, $expected)
+    public function testWakeup($callable, $val, $expected)
     {
-        $callback = new Callback($callable);
-        /* @var $callback Callback */
+        $callback = new SerializedCallback($callable);
+        /* @var $callback SerializedCallback */
         $wakeupedCallback = unserialize(serialize($callback));
         $this->assertEquals($expected, $wakeupedCallback($val));
     }
-
-    /*
-    public function test__wakeupWithPromise($callable, $val, $expected)
-    {
-        $callback = new Callback($callable);
-        $wakeupedCallback = unserialize(serialize($callback));
-        $masterPromise = new Promise();
-        $slavePromise = $masterPromise->then($wakeupedCallback);
-        $masterPromise->resolve($val);
-        $this->assertEquals($expected, $slavePromise->wait(false));
-    }*/
-
 }
