@@ -1,34 +1,30 @@
 <?php
 /**
- * Created by PhpStorm.
- * User: root
- * Date: 21.03.17
- * Time: 15:26
+ * @copyright Copyright © 2014 Rollun LC (http://rollun.com/)
+ * @license LICENSE.md New BSD License
  */
 
 namespace rollun\test\callback\Callback;
 
 use rollun\callback\Callback\Ticker;
-use rollun\installer\Command;
 
 class TickerTest extends CallbackTestDataProvider
 {
     const FILE_WITH_RESULT = 'ticker_results';
+
     /**
+     * @dataProvider providerTickerType()
      * @param $tickerCallback
      * @param $ticksCount
      * @param $tickDuration
      * @param $delayMicroSecond
-     * @param $value
-     * @internal param array $interruptors
-     * @internal param $val
-     * @dataProvider providerTickerType()
      */
     public function test($tickerCallback, $ticksCount, $tickDuration, $delayMicroSecond)
     {
-        $file = Command::getDataDir() . static::FILE_WITH_RESULT;
-        if(file_exists($file)) {
-           unlink($file);
+        $file = 'data/' . static::FILE_WITH_RESULT;
+
+        if (file_exists($file)) {
+            unlink($file);
         }
 
         $sleepTime = ($ticksCount * $tickDuration) + ($delayMicroSecond / 1000000);
@@ -37,7 +33,7 @@ class TickerTest extends CallbackTestDataProvider
         $startTime = microtime(true);
         $ticker($file);
         $finishTime = microtime(true);
-        
+
         $workTimeDiff = abs(($finishTime - $startTime) - $sleepTime);
         $this->assertTrue($workTimeDiff >= 0 && $workTimeDiff <= 0.03);
 
@@ -51,6 +47,10 @@ class TickerTest extends CallbackTestDataProvider
             $this->assertTrue($diff >= 0);
             $this->assertGreaterThanOrEqual(0, $diff);
             $this->assertLessThanOrEqual(0.05, $diff);
+        }
+
+        if (file_exists($file)) {
+            unlink($file);
         }
     }
 }
