@@ -1,15 +1,17 @@
 <?php
+/**
+ * @copyright Copyright © 2014 Rollun LC (http://rollun.com/)
+ * @license LICENSE.md New BSD License
+ */
 
 namespace rollun\test\callback\Queues;
 
-use rollun\callback\Queues\Queue;
-use rollun\promise\Promise\Promise;
+use PHPUnit\Framework\TestCase;
+use rollun\callback\Queues\FileQueue;
 use rollun\callback\Queues\QueueInterface;
-use rollun\dic\InsideConstruct;
 
-class QueueTest extends \PHPUnit_Framework_TestCase
+class QueueTest extends TestCase
 {
-
     /**
      * @var QueueInterface
      */
@@ -17,23 +19,20 @@ class QueueTest extends \PHPUnit_Framework_TestCase
 
     protected function setUp()
     {
-
-        $this->object = new Queue('test_queue');
+        $this->object = new FileQueue('test_queue');
         $this->object->purgeQueue('test_queue');
     }
 
-    public function test__getNullMessage()
+    public function testGetNullMessage()
     {
         $message = $this->object->getMessage();
         $this->assertEquals(null, $message);
     }
 
-    public function test__addMessage()
+    public function testAddMessage()
     {
-
         $this->object->addMessage('test1');
         $message = $this->object->getMessage();
         $this->assertEquals('test1', $message->getData());
     }
-
 }
